Move keyring-files in designated dir and remove keyring-dependency
